+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@
 
     * Loans & Payments
         * [CP-3646] - Allow loan charges to optionally be rounded down to the nearest whole number
+    * External Services
+        * [CP-3554] - Modify payment provider form validation to cater for providers using API key as authentication method
+
 
     * Translations
         * [CP-3366] - Translation in Organisation related UI